% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight_midvariance.R
\name{biweight_midvariance}
\alias{biweight_midvariance}
\title{Biweight Midvariance and the Square Root of the Biweight Midvariance}
\usage{
biweight_midvariance(x, sqrt_bwmv = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{sqrt_bwmv}{A boolean value indicating whether the Square Root of the
Biweight Midvariance should be calculaed. The default is set to TRUE.}
}
\description{
The present function is used to calculate the robust Biweight Midvariance of
a vector of values.
The Square Root of the Biweight Midvariance can also be calculated.
}
\section{Bibliography}{


Rodriguez-Martin, M. et al. (2019) Validation of Portable Mobile Mapping System
for Inspection Tasks
In Thermal and Fluid Mechanical Facilities. Remote Sensing. 11(19):2205-2219

Nocerino, E. et al. (2017)
Investigation of Indoor and Outdoor Performance of Two Portable Mobile Mapping
Systems. Proceedings Volume 10332, Videometrics, Range Imaging, and Applications XIV
DOI: 10.1117/12.2270761

Courtenay, L.A. et al. (2020) Obtaining New Resolutions in Carnivore Tooth Pit Morphological
Analyses: A Methodological Update for Digital Taphonomy. PLoS ONE. 15(10):e0240328

Courtenay, L.A. et al. (2021) Hyperspectral Imaging and Robust Statistics in Non-Melanoma
Skin Cancer Analysis, Biomedical Optics Express. 12(8):5107-5127
}

\examples{

trial_data <- rnorm(100)
biweight_midvariance(trial_data)

}
