% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Plot p-Value Calibration Curves}
\usage{
calibration_curve(
  p_values = c(1, 1e-05),
  method = c("BFB", "FPR", "p_BFB vs FPR", "p_H0"),
  priors = NULL,
  lwd = 1,
  col = "black",
  add = FALSE
)
}
\arguments{
\item{p_values}{A set of two numeric values between 0 and 1 defining the minimum
and maximum p-Values the calibration curve will be plotted for.}

\item{method}{a string defining the p-Value calibration method, select one of
the following options: Bayes Factor Bounds ("BFB"); False Positive Risk ("FPR");
Bayes Factor Bound probability vs False Positive Risk ("p_BFB vs FPR");
Probability of the Null Hypothesis ("p_H0"). By default BFB is selected}

\item{priors}{A numeric value between 0 and 1 defining the prior probability in
favour of the null hypothesis.}

\item{lwd}{Line width of the curve}

\item{col}{Colour of the line}

\item{add}{A boolean value that can be used to add curves to an already existing plot}
}
\description{
The present function can be used to plot calibration curves for a set of p-Values
according to different Bayseian based metrics
}
\examples{

# example 1

calibration_curve(c(1, 0.0001))
calibration_curve(c(0.01, 0.0001), method = "BFB")

# example 2

calibration_curve(method = "FPR", lwd = 2)
calibration_curve(method = "FPR", priors = 0.7,
                  col = "red", lwd = 2, add = TRUE)
calibration_curve(method = "FPR", priors = 0.3,
                  col = "blue", lwd = 2, add = TRUE)
legend("topright", inset = 0.02,
       legend = c("0.7", "0.5", "0.3"), col = c("black", "red", "blue"),
       lty = 1, lwd = 2)

# example 3

par(mfrow = c(1, 3))
calibration_curve(method = "p_BFB vs FPR", priors = 0.7, lwd = 2)
calibration_curve(method = "p_BFB vs FPR", priors = 0.5, lwd = 2)
calibration_curve(method = "p_BFB vs FPR", priors = 0.3, lwd = 2)
par(mfrow = c(1, 1))

# example 4

calibration_curve(method = "p_H0", priors = 0.7, lwd = 2, col = "black")
calibration_curve(method = "p_H0", priors = 0.5, lwd = 2, col = "red", add = TRUE)
calibration_curve(method = "p_H0", priors = 0.3, lwd = 2, col = "blue", add = TRUE)
legend("topright", inset = 0.02,
       legend = c("0.7", "0.5", "0.3"), col = c("black", "red", "blue"),
       lty = 1, lwd = 2)

}
\seealso{
\code{\link{FPR}}
}
