% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_absolute_deviation.R
\name{median_absolute_deviation}
\alias{median_absolute_deviation}
\title{(Normalised) Median Absolute Deviation}
\usage{
median_absolute_deviation(x, nmad = TRUE, constant = 1.4826)
}
\arguments{
\item{x}{A vector of numeric values}

\item{nmad}{A boolean value indicating whether the normalised MAD value
should be computed. The default is set to TRUE.}

\item{constant}{A numeric value defining the value which to normalise MAD values
with}
}
\description{
The present function is used to calculate the Median Absolute Deviation of a vector
of values. By default, this function calculates the normalised version.
}
\section{Bibliography}{


Hohle, J. and Hohle, M. (2009) Accuracy Assessment of Digital Elevation Models by
Means of Robust Statistical Methods, ISPRS Journal of Photogrammetry and Remote Sensing.
64(4):398-406

Rodriguez-Martin, M. et al. (2019) Validation of Portable Mobile Mapping System for Inspection Tasks
In Thermal and Fluid Mechanical Facilities. Remote Sensing. 11(19):2205-2219

Courtenay, L.A. et al. (2020) Obtaining New Resolutions in Carnivore Tooth Pit Morphological
Analyses: A Methodological Update for Digital Taphonomy. PLoS ONE. 15(10):e0240328

Courtenay, L.A. et al. (2021) Hyperspectral Imaging and Robust Statistics in Non-Melanoma
Skin Cancer Analysis, Biomedical Optics Express. 12(8):5107-5127
}

\examples{

trial_data <- rnorm(100)
median_absolute_deviation(trial_data)

}
