% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_CI.R
\name{quantile_CI}
\alias{quantile_CI}
\title{Interquantile Range and Confidence Intervals}
\usage{
quantile_CI(x, q = c(0.025, 0.975))
}
\arguments{
\item{x}{A vector of numeric values}

\item{q}{A set of two numeric values between 0 and 1 that define the confidence
intervals the user wishes to calculate. If only one value is provided then this
function simply returns the specified quantile. Default intervals are set to 95\%}
}
\value{
Either the lower and upper quantile confidence intervals or a specified quantile
}
\description{
The present function is used to calculate asymmetric measurements of
sample dispersal using the Interquantile range.
}
\section{Bibliography}{


Hohle, J. and Hohle, M. (2009) Accuracy Assessment of Digital Elevation Models by
Means of Robust Statistical Methods, ISPRS Journal of Photogrammetry and Remote Sensing.
64(4):398-406
}

\examples{

trial_data <- rnorm(100)
quantile_CI(trial_data)

}
